backend/requirements.txt
# DHsystem MCP Backend - Python Dependencies
# Cài đặt: pip install -r requirements.txt

# ── Web Framework ───────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── MCP Protocol (Official SDK + Xiaozhi Bridge) ─────────────
mcp>=1.20.0
fastmcp>=2.13.0
mcp-proxy>=0.10.0

# ── Database: Appwrite Cloud ─────────────────────────────────
# Thay thế Supabase - dễ dùng hơn, UI tốt hơn
appwrite>=6.0.0

# ── Vector Store: ChromaDB (local, zero-config) ───────────────
# Chạy local, persist vào ./chroma_data/
chromadb>=0.6.0

# ── AI: Google Gemini (mới nhất, không deprecated) ───────────
google-genai>=1.10.0

# ── WebSocket Server (cho ESP32 kết nối trực tiếp) ───────────
websockets>=12.0

# ── HTTP Client ──────────────────────────────────────────────
httpx>=0.27.0
aiohttp>=3.9.0

# ── PDF Support (tùy chọn, để ingest file PDF) ───────────────
# pypdf>=4.0.0

# ── Audio Codec: Opus <-> PCM (tùy chọn) ─────────────────────
# Nếu không cài, dùng fallback ffmpeg (cần cài ffmpeg riêng)
# Windows: tải libopus tại https://opus-codec.org/downloads/
# opuslib>=3.0.1

# ── Utilities ────────────────────────────────────────────────
python-multipart>=0.0.9
structlog>=24.1.0
